Moving get_entities() and allocate_ids() to designated module.
This is a pre-cursor to designated put(), delete() and get()
methods in this module (inspired by #421).
